% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{build_predicted_pmf}
\alias{build_predicted_pmf}
\title{Build predicted PMF}
\usage{
build_predicted_pmf(i, baseline_pmf, eigenval, scores, eigenfun)
}
\arguments{
\item{i}{integer, indivdiual index.}

\item{baseline_pmf}{numerical vector containing the values for the
baseline for PMF prediction.}

\item{eigenval}{numerical vector containing eigenvalues for each principal
component.}

\item{scores}{data frame of individual scores (in rows) for each principal
component (in columns).}

\item{eigenfun}{data frame containing the values of the eigenfunctions for
each principal components (in columns) over the occurrence \code{grid}
(in rows).}
}
\value{
numerical vector containing the values for the predicted PMF
function.
}
\description{
Build individual predicted PMF from Karhunen–Loève
decomposition
}
\details{
The Principal Component Analysis for Point Process data is based on the
Karhunen–Loève decomposition of the empirical \emph{cumulative mass functions}.

See Picard et al (2024) or the tutorial vignette "PCA for Point Processes
with Earthquake Data" for more details:
\code{vignette("01-introduction-pppca", package = "pppca")}
}
\keyword{internal}
