% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empirical.R
\name{get_empirical_pmf}
\alias{get_empirical_pmf}
\title{Empirical cumulative probability mass function}
\usage{
get_empirical_pmf(PP, mc.cores = parallelly::availableCores())
}
\arguments{
\item{PP}{list of \strong{numerical} vectors containing occurrences of each
individual realization of the point process.}

\item{mc.cores}{The number of cores to use, i.e.\ifelse{latex}{\out{~}}{ }{}at most how many
    child processes will be run simultaneously.   The option is
    initialized from environment variable \env{MC_CORES} if set.  Must
    be at least one, and parallelization requires at least two cores.}
}
\value{
List of empirical cumulative PMF values for each individual point
process realization passed in input.
}
\description{
Compute the empirical cumulative mass function (PMF) for each realization of
the point process.
}
\details{
The empirical cumulative PMF values for each indidividual point process
realization are computed along the union of all individual point process
occurrences.

See Picard et al (2024) or the tutorial vignette "PCA for Point Processes
with Earthquake Data" for more details:
\code{vignette("01-introduction-pppca", package = "pppca")}
}
\examples{
emp_pmf <- get_empirical_pmf(earthquake)
}
\references{
Picard, Franck, Vincent Rivoirard, Angelina Roche, and Victor
Panaretos. 2024. “PCA for Point Processes.” arXiv.
\doi{10.48550/arXiv.2404.19661}.
}
\seealso{
\code{\link[=get_empirical_intensity]{get_empirical_intensity()}}
}
