% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicted.R
\name{get_predicted_pmf}
\alias{get_predicted_pmf}
\title{Predicted cumulative probability mass function}
\usage{
get_predicted_pmf(
  PP,
  eigenval,
  scores,
  eigenfun,
  mc.cores = parallelly::availableCores()
)
}
\arguments{
\item{PP}{list of \strong{numerical} vectors containing occurrences of each
individual realization of the point process.}

\item{eigenval}{numerical vector containing eigenvalues for each principal
component.}

\item{scores}{data frame of individual scores (in rows) for each principal
component (in columns).}

\item{eigenfun}{data frame containing the values of the eigenfunctions for
each principal components (in columns) over the occurrence \code{grid}
(in rows).}

\item{mc.cores}{The number of cores to use, i.e.\ifelse{latex}{\out{~}}{ }{}at most how many
    child processes will be run simultaneously.   The option is
    initialized from environment variable \env{MC_CORES} if set.  Must
    be at least one, and parallelization requires at least two cores.}
}
\value{
a matrix containing the values of the predicted PMF for
each individual realization of the point process (in columns) over the
union of all occurrences of all point process realizations (in rows).
}
\description{
Compute the predicted cumulative mass function (PMF) for each realization
of the point process using the Karhunen–Loève decomposition associated
with the PPPCA decomposition.
}
\details{
The predicted cumulative PMF values for each indidividual point process
realization are computed along the union of all individual point process
occurrences.

See Picard et al (2024) or the tutorial vignette "PCA for Point Processes
with Earthquake Data" for more details:
\code{vignette("01-introduction-pppca", package = "pppca")}
}
\examples{
# conversion of dates to numerical values
earthquake_num <- lapply(earthquake, as.numeric)
# apply principal component analysis to point process data
pppca_res <- pppca(earthquake_num, Jmax = 50, mc.cores = 1)
# reconstruct predicted PMF
pred_pmf <- get_predicted_pmf(
    earthquake, pppca_res$eigenval, pppca_res$scores, 
    pppca_res$eigenfun, mc.cores = 1
)
}
\references{
Picard, Franck, Vincent Rivoirard, Angelina Roche, and Victor
Panaretos. 2024. “PCA for Point Processes.” arXiv.
\doi{10.48550/arXiv.2404.19661}.
}
\seealso{
\code{\link[=pppca]{pppca()}}, \code{\link[=get_predicted_intensity]{get_predicted_intensity()}}
}
