% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pppca-package.R
\docType{package}
\name{pppca-package}
\alias{pppca-package}
\title{Point Process Principal Component Analysis}
\description{
Principal Component Analysis (PCA) for dimension reduction and variance
decomposition of point process data
}
\details{
The Principal Component Analysis for Point Process data is based on the
Karhunen–Loève decomposition of the empirical \emph{cumulative mass functions}.

See Picard et al (2024) or the tutorial vignette "PCA for Point Processes
with Earthquake Data" for more details:
\code{vignette("01-introduction-pppca", package = "pppca")}
}
\references{
Picard, Franck, Vincent Rivoirard, Angelina Roche, and Victor
Panaretos. 2024. “PCA for Point Processes.” arXiv.
\doi{10.48550/arXiv.2404.19661}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/franckpicard/pppca}
}

}
\author{
\itemize{
\item Franck Picard (\href{https://franckpicard.github.io/}{link}|
\href{https://orcid.org/0000-0001-8084-5481}{ORCID})
\item Ghislain Durif (\href{https://gdurif.perso.math.cnrs.fr/}{link}|
\href{https://orcid.org/0000-0003-2567-1401}{ORCID})
}
}
