% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pppca.R
\name{pppca}
\alias{pppca}
\title{Point Process Principal Component Analysis (PPPCA)}
\usage{
pppca(PP, Jmax, mc.cores = parallelly::availableCores())
}
\arguments{
\item{PP}{list of \strong{numerical} vectors containing occurrences of each
individual realization of the point process.}

\item{Jmax}{integer, number of principal components to consider.}

\item{mc.cores}{The number of cores to use, i.e.\ifelse{latex}{\out{~}}{ }{}at most how many
    child processes will be run simultaneously.   The option is
    initialized from environment variable \env{MC_CORES} if set.  Must
    be at least one, and parallelization requires at least two cores.}
}
\value{
PCA decomposition as a list containing:
\itemize{
\item \code{grid}: numerical vector of the union of all occurrences of all point
process realizations in \code{PP} input argument.
\item \code{eigenval}: numerical vector of length \code{Jmax} containing eigenvalues for
each principal component.
\item \code{percentvar}: numerical vector of length \code{Jmax} containing the percentage
of variance in the data explained by each principal component.
\item \code{eigenfun}: data frame containing the values of the eigenfunctions for
each principal components (in columns) over the occurrence \code{grid}
(in rows).
\item \code{scores}: data frame of individual scores (in rows) for each
principal component (in columns).
}
}
\description{
Principal Component Analysis for Point Process data.
}
\details{
The Principal Component Analysis for Point Process data is based on the
Karhunen–Loève decomposition of the empirical \emph{cumulative mass functions}.

See Picard et al (2024) or the tutorial vignette "PCA for Point Processes
with Earthquake Data" for more details:
\code{vignette("01-introduction-pppca", package = "pppca")}

\strong{Important:} if the point process occurrences are not numerical values
(e.g. dates), they should be converted to numerical values before using
this function.
}
\examples{
# conversion of dates to numerical values
earthquake_num <- lapply(earthquake, as.numeric)
# apply principal component analysis to point process data
pppca_res <- pppca(earthquake_num, Jmax = 50, mc.cores = 1)
}
\references{
Picard, Franck, Vincent Rivoirard, Angelina Roche, and Victor
Panaretos. 2024. “PCA for Point Processes.” arXiv.
\doi{10.48550/arXiv.2404.19661}.
}
